(fuji-testnet): add production-preview key
