+++ src/components/Header/AppHeaderUser.tsx
@@ -7,7 +7,6 @@ import connectWalletImg from "img/ic_wallet_24.svg";
 
 import "./Header.css";
 import { isHomeSite, getAccountUrl } from "lib/legacy";
-import { isDevelopment } from "lib/legacy";
 import cx from "classnames";
 import { Trans } from "@lingui/macro";
 import NetworkDropdown from "../NetworkDropdown/NetworkDropdown";
@@ -15,6 +14,7 @@ import LanguagePopupHome from "../NetworkDropdown/LanguagePopupHome";
 import { ARBITRUM, ARBITRUM_TESTNET, AVALANCHE, AVALANCHE_FUJI_TESTNET, getChainName } from "config/chains";
 import { switchNetwork } from "lib/wallets";
 import { useChainId } from "lib/chains";
+import { isDevelopment } from "config/env";
 
 type Props = {
   openSettings: () => void;
