+++ src/config/localStorage.ts
@@ -12,3 +12,5 @@ export const DISABLE_ORDER_VALIDATION_KEY = "disable-order-validation";
 export const SHOULD_SHOW_POSITION_LINES_KEY = "Exchange-swap-should-show-position-lines";
 export const REFERRAL_CODE_KEY = "GMX-referralCode";
 export const REFERRALS_SELECTED_TAB_KEY = "Referrals-selected-tab";
+export const REDIRECT_POPUP_TIMESTAMP_KEY = "redirect-popup-timestamp";
+export const PRODUCTION_PREVIEW_KEY = "production-preview";
